test functions moved to new module
